test-requirements.txt
hacking>=0.9.2,<0.10

coverage>=3.6
discover
fixtures>=1.3.1
mock>=1.2;python_version!='2.6'
mock==1.0.1;python_version=='2.6'
oslosphinx
oslotest
os-testr>=0.4.1
sphinx>=1.1.2,!=1.2.0,!=1.3b1,<1.3
sphinxcontrib-pecanwsme>=0.8
sphinxcontrib-httpdomain
stevedore>=0.14
testrepository>=0.0.18
testscenarios>=0.4
testtools>=1.4.0
tempest-lib>=0.12.0
